Address PR #181 review feedback
